Use InfomaniakCore classes
