+++ Project.swift
@@ -22,7 +22,7 @@ import ProjectDescriptionHelpers
 let project = Project(name: "kDrive",
                       packages: [
                           .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "5.2.2")),
-                          .package(url: "https://github.com/Infomaniak/ios-core.git", .upToNextMajor(from: "1.1.2")),
+                          .package(url: "https://github.com/Infomaniak/ios-core.git", .upToNextMajor(from: "1.1.7")),
                           .package(url: "https://github.com/Infomaniak/ios-login.git", .upToNextMajor(from: "1.4.0")),
                           .package(url: "https://github.com/realm/realm-cocoa", .upToNextMajor(from: "10.0.0")),
                           .package(url: "https://github.com/SCENEE/FloatingPanel", .upToNextMajor(from: "2.0.0")),
