+++ kDrive/UI/Controller/Files/File List/FileListViewModel.swift
@@ -19,6 +19,7 @@
 import CocoaLumberjackSwift
 import Combine
 import Foundation
+import InfomaniakCore
 import kDriveCore
 import kDriveResources
 import RealmSwift
