+++ kDrive/UI/Controller/Files/FileActionsFloatingPanelViewController.swift
@@ -17,6 +17,7 @@
  */
 
 import CocoaLumberjackSwift
+import InfomaniakCore
 import kDriveCore
 import kDriveResources
 import Sentry
