+++ kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.swift
@@ -16,6 +16,7 @@
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import InfomaniakCore
 import kDriveCore
 import kDriveResources
 import UIKit
