+++ kDriveCore/Data/MQService/MQService.swift
@@ -18,6 +18,7 @@
 
 import CocoaLumberjackSwift
 import Foundation
+import InfomaniakCore
 import MQTTNIO
 
 public class MQService {
