+++ kDriveCore/Utils/NotificationsHelper.swift
@@ -158,7 +158,7 @@ public enum NotificationsHelper {
                 return
             }
 
-            let isInBackground = Constants.isInExtension || UIApplication.shared.applicationState != .active
+            let isInBackground = Bundle.main.isExtension || UIApplication.shared.applicationState != .active
 
             if isInBackground {
                 let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 0.1, repeats: false)
